requirements.txt
Django
psycopg2-binary
djangorestframework
python-decouple
Pillow
django-filter
django-user-accounts